Improve printf handling of size_t args
The %*s format takes two arguments, a precision length/width and
an actual string; the length is specified as a signed integer.
The size_t length field of the gss_buffer_desc type is an unsigned
type, which must be cast or otherwise converted to a signed type
to match the format string expectations.
I do not think that the length will approach SIZE_T_MAX in practice,
due to buffer constraints, so do not include handling for the
edge case.

There is a '%zu' format string for printing size_ts, but it is not
available everywhere (e.g., AIX).  Instead, use the
unsigned long long abomination.

+++ src/appl/gss-sample/gss-server.c
@@ -893,7 +893,7 @@ showLocalIdentity(OM_uint32 *minor, gss_name_t name)
 
     major = gss_localname(minor, name, GSS_C_NO_OID, &localname);
     if (major == GSS_S_COMPLETE)
-        printf("localname: %-*s\n", localname.length, localname.value);
+        printf("localname: %-*s\n", (int)localname.length, localname.value);
     else if (major != GSS_S_UNAVAILABLE)
         display_status("gss_localname", major, *minor);
     gss_release_buffer(minor, &localname);
